rsx/common/d3d12: Move surface_store in common
